+++ tests/testthat/test-column-type.R
@@ -90,6 +90,23 @@ test_that("Character columns are character - SPSS", {
   expect_is(NIBRS$JUDICIAL_DISTRICT, "character")
 
   expect_is(weimar[, 4], "character")
+
+  expect_is(SHR$NUMERIC_STATE_CODE, "character")
+  expect_is(SHR$SITUATION, "character")
+  expect_is(SHR$VICTIM_4_ETHNIC_ORIGIN, "character")
+  expect_is(SHR$VICTIM_1_SEX, "character")
+  expect_is(SHR$OFFENDER_1_WEAPON, "character")
+
+  expect_is(UCR$ID_CODE, "character")
+  expect_is(UCR$NUMERIC_STATE_CODE, "character")
+  expect_is(UCR$GROUP_NUMBER, "character")
+  expect_is(UCR$DIVISION, "character")
+  expect_is(UCR$NUMBER_OF_MONTHS_REPORTED, "character")
+
+  expect_is(NIBRS$SEGMENT_LEVEL, "character")
+  expect_is(NIBRS$NUMERIC_STATE_CODE, "character")
+  expect_is(NIBRS$CORE_CITY, "character")
+  expect_is(NIBRS$POPULATION_GROUP, "character")
 })
 
 test_that("Numeric columns are Numeric - SPSS", {
@@ -120,26 +137,27 @@ test_that("Numeric columns are Numeric - SPSS", {
   expect_is(weimar[, 11], "numeric")
 })
 
-test_that("Factor columns are Factor - SPSS", {
-  expect_is(SHR$NUMERIC_STATE_CODE, "factor")
-  expect_is(SHR$SITUATION, "factor")
-  expect_is(SHR$VICTIM_4_ETHNIC_ORIGIN, "factor")
-  expect_is(SHR$VICTIM_1_SEX, "factor")
-  expect_is(SHR$OFFENDER_1_WEAPON, "factor")
 
-  expect_is(UCR$ID_CODE, "factor")
-  expect_is(UCR$NUMERIC_STATE_CODE, "factor")
-  expect_is(UCR$GROUP_NUMBER, "factor")
-  expect_is(UCR$DIVISION, "factor")
-  expect_is(UCR$NUMBER_OF_MONTHS_REPORTED, "factor")
-
-  expect_is(NIBRS$SEGMENT_LEVEL, "factor")
-  expect_is(NIBRS$NUMERIC_STATE_CODE, "factor")
-  expect_is(NIBRS$CORE_CITY, "factor")
-  expect_is(NIBRS$POPULATION_GROUP, "factor")
-})
 
 test_that("Character columns are character - SAS", {
+
+  expect_is(SHR_sas$NUMERIC_STATE_CODE, "character")
+  expect_is(SHR_sas$SITUATION, "character")
+  expect_is(SHR_sas$VICTIM_4_ETHNIC_ORIGIN, "character")
+  expect_is(SHR_sas$VICTIM_1_SEX, "character")
+  expect_is(SHR_sas$OFFENDER_1_WEAPON, "character")
+
+  expect_is(UCR_sas$ID_CODE, "character")
+  expect_is(UCR_sas$NUMERIC_STATE_CODE, "character")
+  expect_is(UCR_sas$GROUP_NUMBER, "character")
+  expect_is(UCR_sas$DIVISION, "character")
+  expect_is(UCR_sas$NUMBER_OF_MONTHS_REPORTED, "character")
+
+  expect_is(NIBRS_sas$SEGMENT_LEVEL, "character")
+  expect_is(NIBRS_sas$NUMERIC_STATE_CODE, "character")
+  expect_is(NIBRS_sas$CORE_CITY, "character")
+  expect_is(NIBRS_sas$POPULATION_GROUP, "character")
+
   expect_is(SHR_sas$ORI_CODE, "character")
 
   expect_is(UCR_sas$ORI_CODE, "character")
@@ -184,21 +202,4 @@ test_that("Numeric columns are Numeric - SAS", {
   expect_is(weimar_sas[, 11], "numeric")
 })
 
-test_that("Factor columns are Factor - SAS", {
-  expect_is(SHR_sas$NUMERIC_STATE_CODE, "factor")
-  expect_is(SHR_sas$SITUATION, "factor")
-  expect_is(SHR_sas$VICTIM_4_ETHNIC_ORIGIN, "factor")
-  expect_is(SHR_sas$VICTIM_1_SEX, "factor")
-  expect_is(SHR_sas$OFFENDER_1_WEAPON, "factor")
 
-  expect_is(UCR_sas$ID_CODE, "factor")
-  expect_is(UCR_sas$NUMERIC_STATE_CODE, "factor")
-  expect_is(UCR_sas$GROUP_NUMBER, "factor")
-  expect_is(UCR_sas$DIVISION, "factor")
-  expect_is(UCR_sas$NUMBER_OF_MONTHS_REPORTED, "factor")
-
-  expect_is(NIBRS_sas$SEGMENT_LEVEL, "factor")
-  expect_is(NIBRS_sas$NUMERIC_STATE_CODE, "factor")
-  expect_is(NIBRS_sas$CORE_CITY, "factor")
-  expect_is(NIBRS_sas$POPULATION_GROUP, "factor")
-})