+++ NEWS.md
@@ -1,7 +1,8 @@
 # asciiSetupReader
 
 # 1.4.0
-+ Adds capacity to the SPSS reader to handle Missing Values
++ Adds capacity to the SPSS reader to handle Missing Values    
 + Fixed bug where function fails if no value-labels are detected
-    + Applied both when setup file has no value-labels and if only select columns that have no value-labels
-+ Added a huge number of tests
+    + Applied both when setup file has no value-labels and if only select columns that have no value-labels    
++ Added a huge number of tests    
++ Changes factor column to character columns